% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.fit.R
\name{beta.fit}
\alias{beta.fit}
\title{Perform Beta Regression for DMP Analysis}
\usage{
beta.fit(beta_matrix, sample_groups, chunk_size = 100, num_cores = 2)
}
\arguments{
\item{beta_matrix}{A matrix of beta values with CpG sites as rows and samples as columns.}

\item{sample_groups}{A vector indicating the phenotype group for each sample. First level should represent controls and second level should represent cases}

\item{chunk_size}{An integer specifying the number of CpG sites to process in each chunk. Default is 100.}

\item{num_cores}{An integer specifying the number of cores to use for parallel processing. Default is 2.}
}
\value{
A dataframe containing regression results for each CpG site.
}
\description{
This function fits a beta regression model to each CpG site in the beta value matrix
and returns a dataframe with regression results, including log fold change and adjusted p-values.
}
\examples{
\dontrun{
results_df <- beta.fit(myNorm, myPD$Sample_Group, chunk_size = 100, num_cores = 10)
}
}
