% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.DMPs.R
\name{filter.DMPs}
\alias{filter.DMPs}
\title{Filter Differentially Methylated Positions (DMPs)}
\usage{
filter.DMPs(
  results_df,
  filter_by = "logFC",
  logFC_threshold = 1,
  meanDiff_threshold = 0.1,
  pval_adj_threshold = 0.05
)
}
\arguments{
\item{results_df}{A dataframe containing regression results, including logFC and adjusted p-values.}

\item{filter_by}{Option to choose "logFC" or "MeanDiff" to filter DMPs. Default is "logFC"}

\item{logFC_threshold}{A numeric value specifying the minimum absolute log fold change for significance. Used when filter_by is set to "logFC". Default is 1.}

\item{meanDiff_threshold}{A numeric value specifying the minimum absolute mean difference for significance. Used when filter_by is set to "MeanDiff". Default is 0.1.}

\item{pval_adj_threshold}{A numeric value specifying the maximum adjusted p-value for significance. Default is 0.05.}
}
\value{
A dataframe containing only the DMPs that meet the specified criteria.
}
\description{
This function filters the results dataframe to identify significant DMPs based on
specified thresholds for log fold change (logFC) and adjusted p-value.
}
\examples{
\dontrun{
filtered_dmps <- filter.DMPs(results_df, filter_by = "logFC", logFC_threshold = 1, pval_adj_threshold = 0.05)
filtered_dmps <- filter.DMPs(results_df, filter_by = "MeanDiff", meanDiff_threshold = 0.1, pval_adj_threshold = 0.05)
}
}
