% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpg_boxplot.R
\name{plot_cpg_boxplot}
\alias{plot_cpg_boxplot}
\title{Plot Methylation Levels for a Specific CpG Site}
\usage{
plot_cpg_boxplot(beta_matrix, sample_groups, cpg_id)
}
\arguments{
\item{beta_matrix}{A matrix of beta values with CpG sites as rows and samples as columns.}

\item{sample_groups}{A vector indicating the phenotype group for each sample.}

\item{cpg_id}{The name of the CpG site to plot.}
}
\value{
A ggplot object representing the boxplot of methylation levels.
}
\description{
This function creates a boxplot of methylation values for a specific CpG site,
comparing the values across two phenotype groups.
}
\examples{
\dontrun{
plot_cpg_boxplot(myNorm, myPD$Sample_Group, "cg18478105")
}
}
